refactor(#33): move utils to files
